Crate nova_core

Re-exports

pub use config::EnvConfigSource;

pub use config::JsonFileConfigSource;

pub use config::NovaConfig;

pub use config::NovaConfigBuilder;

pub use config::NovaConfigSource;

pub use config::NovaSecretSource;

pub use config::ReloadableConfig;

pub use config::spawn_json_file_hot_reloader;

pub use error::ErrorResponse;

pub use error::NovaError;

pub use error::NovaResult;

pub use discovery::Discovery;

pub use discovery::InstanceStatus;

pub use discovery::ServiceInstance;

pub use error::DiscoveryError;

pub use runtime::NovaApp;

pub use runtime::NovaRoute;

pub use traits::NovaLifecycle;

pub use traits::NovaModule;

pub use traits::NovaPlugin;

pub use traits::NovaRequestModel;

pub use traits::NovaResponseModel;

pub use traits::NovaRouterExtender;

pub use state::NovaState;

pub use axum;

pub use inventory;

Modules

config

discovery

error

runtime

state

traits

Structs

Json

JSON Extractor / Response.

Enums

LuaValue

Lua script return values supported by resilience stores.

Traits

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

ResilienceStore

Minimal distributed key-value operations used by resilience backends.

Serialize

A data structure that can be serialized into any data format supported by Serde.

Attribute Macros

async_trait

delete

Attach a DELETE route to a free function. Usage: #[delete("/path")].

get

Attach a GET route to a free function. Usage: #[get("/path")].

patch

Attach a PATCH route to a free function. Usage: #[patch("/path")].

post

Attach a POST route to a free function. Usage: #[post("/path")].

put

Attach a PUT route to a free function. Usage: #[put("/path")].

rest_controller

Marks a struct as a Nova controller. The macro currently adds a nova_metadata() helper; route handlers are attached via get/post/... attributes on impl functions.

Derive Macros

Deserialize

NovaRequest

Derive macro to mark a type as a Nova request model.

NovaResponse

Derive macro to mark a type as a Nova response model.

Serialize